requirements.txt
pyspark==3.5.0
kafka-python
pyyaml